# IntPay.Api/docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: IntPay API
  version: 1.0.0
  description: |
    OpenAPI documentation for IntPay API v1.

    The current implementation uses explicit caller identifiers (`userId`, `profileId`, or `actingUserId`) as a temporary request-level authorization boundary. JWT/session authentication is not enforced by the API yet.

    In the current MVP, authorized card GET responses may include the full card PAN (`cardNumber`) and CVV (`cvv`) for demo flows. Production deployments should re-enable masking, secret handling, and real authentication.

    `useTimes` is optional when creating an intent. If the client omits it or sends `null`, the API treats the intent as unlimited. The current numeric sentinel for unlimited usage is `99999`; API consumers should display this as `unlimited`.

    In Development, `MapOpenApi()` still exposes generated JSON (e.g. `/openapi/v1.json`). Scalar is configured to load this hand-maintained YAML from `/openapi/intpay-v1.yaml` (same contents as `docs/openapi-v1.yaml` on disk) so examples and long descriptions appear in the UI. Keep this file aligned with `Endpoints/V1` when behavior changes.
servers:
  - url: /api/v1
    description: API v1 relative server
...
paths:
...
  /cards/{cardId}:
    get:
      tags: [Cards]
      summary: Get card with audit logs
      description: Returns a rich card response and audit logs for a single card. The caller must be the creator or receiver.
      operationId: getCardWithLogs
      parameters:
        - $ref: '#/components/parameters/CardIdPath'
        - name: profileId
          in: query
          required: true
          description: Caller profile id. Must be a positive integer and must belong to the card creator or receiver.
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 1
      responses:
        '200':
          description: Card fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardWithLogsEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: Card fetched successfully.
                    data:
                      card:
                        intentId: 21
                        city: Riyadh
                        country: SA
                        mccList: []
                        requiredInvoiceProve: true
                        card:
                          id: 31
                          cardNumber: '4111123412341234'
                          last4: '1234'
                          cvv: '123'
                          useTimes: 99999
                          usesLeft: 99999
                          isSpendBlocked: true
                      logs: []
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cards/{cardId}/logs:
    get:
      tags: [Cards]
      summary: Get card audit logs
      description: Returns paged audit logs for a card with optional decision and date filters.
      operationId: getAuditLogsByCardId
      parameters:
        - $ref: '#/components/parameters/CardIdPath'
        - $ref: '#/components/parameters/ActingUserIdRequired'
        - $ref: '#/components/parameters/LimitOptional'
        - $ref: '#/components/parameters/OffsetOptional'
        - name: decision
          in: query
          required: false
          description: Optional audit decision filter.
          schema:
            $ref: '#/components/schemas/DecisionStatus'
          example: approved
        - name: from
          in: query
          required: false
          description: Inclusive created-at lower bound.
          schema:
            type: string
            format: date-time
          example: '2026-05-01T00:00:00Z'
        - name: to
          in: query
          required: false
          description: Inclusive created-at upper bound.
          schema:
            type: string
            format: date-time
          example: '2026-05-02T00:00:00Z'
      responses:
        '200':
          description: Audit logs fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAuditLogsEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: Audit logs fetched successfully.
                    data:
                      cardId: 31
                      total: 1
                      limit: 20
                      offset: 0
                      logs:
                        - id: 90
                          cardId: 31
                          transactionAmount: 25.5
                          merchantName: Riyadh Taxi
                          mcc: '4121'
                          decision: approved
                          reason: null
                          createdAt: '2026-05-01T02:00:00Z'
                          city: Riyadh
                          entityId: 21
                          action: authorization
                          occurredAt: '2026-05-01T02:00:00Z'
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cards/{cardId}/lock-state:
    post:
      tags: [Cards]
      summary: Set card manual freeze state
      description: Toggles manual freeze only. This endpoint does not change invoice-pending lock state.
      operationId: setCardLockState
      parameters:
        - $ref: '#/components/parameters/CardIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCardLockRequest'
            examples:
              freeze:
                value:
                  locked: true
                  actingUserId: 1
              unfreeze:
                value:
                  locked: false
                  actingUserId: 1
      responses:
        '200':
          description: Card manual freeze state updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardManualFreezeEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: Card manual freeze state updated successfully.
                    data:
                      cardId: 31
                      isManuallyFrozen: true
                      isLockedByPendingInvoice: false
                      isSpendBlocked: true
                      previousManualFreeze: false
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cards/by-user/{userId}/latest:
    get:
      tags: [Cards]
      summary: Get latest card for user
      description: Returns the latest card where the user is either creator or receiver.
      operationId: getLatestCardForUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Latest card fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentWithCardEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: Latest card fetched successfully.
                    data:
                      intentId: 21
                      city: Riyadh
                      country: SA
                      mccList: []
                      requiredInvoiceProve: false
                      card:
                        id: 31
                        cardNumber: '4111123412341234'
                        cvv: '123'
                        useTimes: 99999
                        usesLeft: 99999
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cards/by-user/{userId}:
    get:
      tags: [Cards]
      summary: Get cards for user
      description: Returns paged cards where the user is creator or receiver.
      operationId: getCardsForUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/LimitOptional'
        - $ref: '#/components/parameters/OffsetOptional'
      responses:
        '200':
          description: Cards fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCardsEnvelope'
              examples:
                success:
                  value:
                    success: true
                    message: Cards fetched successfully.
                    data:
                      total: 1
                      limit: 10
                      offset: 0
                      items:
                        - intentId: 21
                          city: Riyadh
                          country: SA
                          mccList: []
                          requiredInvoiceProve: false
                          card:
                            id: 31
                            last4: '1234'
                            useTimes: 99999
                            usesLeft: 99999
                    meta:
                      statusCode: 200
                      version: v1
                      timestamp: '2026-05-01T02:00:00.0000000+00:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    UserIdPath:
      name: userId
      in: path
      required: true
      description: User profile id.
      schema:
        type: integer
        format: int32
        minimum: 1
      example: 1
...
    CardIdPath:
      name: cardId
      in: path
      required: true
      description: Virtual card id.
      schema:
        type: integer
        format: int32
        minimum: 1
      example: 31
    ActingUserIdRequired:
      name: actingUserId
      in: query
      required: true
      description: Temporary caller id. Must be a positive profile id and must be authorized for the resource.
      schema:
        type: integer
        format: int32
        minimum: 1
      example: 1
    LimitOptional:
      name: limit
      in: query
      required: false
      description: Page size. Service clamps supported typed endpoints to 1..1000. Some card endpoints fallback to 50 when empty or invalid.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 1000
        default: 50
      example: 20
    OffsetOptional:
      name: offset
      in: query
      required: false
      description: Number of records to skip. Negative values are normalized to 0 on typed endpoints.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
      example: 0
  responses:
    BadRequest:
      description: Invalid request body, invalid query/path parameter, validation failure, or business rule error.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ProblemDetails'
          examples:
            error:
              value:
                success: false
                message: Invalid query parameter value.
                detail: Failed to bind parameter "Nullable<int> limit" from "abc".
            problem:
              value:
                type: about:blank
                title: Bad Request
                status: 400
                detail: Request body is required.
    Unauthorized:
      description: Authentication is not implemented yet, but this standard response is reserved for future JWT/session authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                success: false
                message: Authentication is required.
    Forbidden:
      description: The caller id is not allowed to access the requested resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                success: false
                message: "Access denied: this resource is only available to the sender or recipient."
    NotFound:
      description: Requested profile, intent, card, or related resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                success: false
                message: Resource not found.
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            error:
              value:
                success: false
                message: Unexpected server error.
  schemas:
    DecisionStatus:
      type: string
      description: |
        Audit decision status.
        - `approved`: authorization or verification was approved.
        - `declined`: authorization or verification was declined.
        - `info`: non-financial informational event.
      enum: [approved, declined, info]
      example: approved
...
    CardPerspectiveType:
      type: string
      description: Card `type` as returned on `IntentWithCardResponse.card` (current-user perspective).
      enum: [self, sent, receiver]
      example: sent
...
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          description: Always false for this error shape.
          example: false
        message:
          type: string
          description: Human-readable error message.
          example: Query parameter actingUserId is required.
        detail:
          type: string
          nullable: true
          description: Optional technical detail, especially for binding failures.
          example: Failed to bind parameter "Nullable<int> limit" from "abc".
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request body is required.
    Meta:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        version:
          type: string
          example: v1
        timestamp:
          type: string
          format: date-time
          example: '2026-05-01T02:00:00.0000000+00:00'
...
    SetCardLockRequest:
      type: object
      required: [locked, actingUserId]
      properties:
        locked:
          type: boolean
          description: True to manually freeze the card, false to unfreeze it.
          example: true
        actingUserId:
          type: integer
          format: int32
          description: Caller profile id. Must be creator or receiver.
          minimum: 1
          example: 1
...
    IntentWithCardEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/IntentWithCardResponse'
            meta:
              $ref: '#/components/schemas/Meta'
...
    CardWithLogsEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/CardWithLogsResponse'
            meta:
              $ref: '#/components/schemas/Meta'
    PagedAuditLogsEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/PagedAuditLogsResponse'
            meta:
              $ref: '#/components/schemas/Meta'
    CardManualFreezeEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/CardManualFreezeStateResponse'
            meta:
              $ref: '#/components/schemas/Meta'
    PagedCardsEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data, meta]
          properties:
            data:
              $ref: '#/components/schemas/PagedCardsResponse'
            meta:
              $ref: '#/components/schemas/Meta'
...
    SuccessEnvelopeBase:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
...
    IntentWithCardResponse:
      type: object
      required: [intentId, city, country, mccList, requiredInvoiceProve, card]
      properties:
        intentId:
          type: integer
          example: 21
        city:
          type: string
          example: Riyadh
        country:
          type: string
          example: SA
        description:
          type: string
          nullable: true
          example: Team travel allowance
        mccList:
          type: array
          items:
            $ref: '#/components/schemas/MccItem'
        requiredInvoiceProve:
          type: boolean
          example: true
        card:
          $ref: '#/components/schemas/CardDetailsResponse'
    CardDetailsResponse:
      type: object
      required:
        - id
        - stripeId
        - createdAt
        - status
        - isLockedByPendingInvoice
        - isManuallyFrozen
        - isSpendBlocked
        - cardNumber
        - last4
        - cvv
        - expiryDate
        - expiryMonth
        - expiryYear
        - cardholderName
        - amount
        - remainingAmount
        - useTimes
        - usesLeft
        - minutesToUnlock
        - hoursToUnlock
        - daysToUnlock
        - timeRemainingLeveled
        - daysLocked
        - creatorId
        - retrieveId
        - type
      properties:
        id:
          type: integer
          example: 31
        stripeId:
          type: string
          example: ic_1234567890abcdef1234
        createdAt:
          type: string
          format: date-time
          example: '2026-05-01T02:00:00Z'
        status:
          type: string
          description: Card status value from the database.
          example: active
        isLockedByPendingInvoice:
          type: boolean
          example: false
        isManuallyFrozen:
          type: boolean
          example: false
        isSpendBlocked:
          type: boolean
          description: True when invoice lock or manual freeze blocks spending.
          example: false
        cardNumber:
          type: string
          description: Full card PAN in the current MVP.
          example: '4111123412341234'
        last4:
          type: string
          example: '1234'
        cvv:
          type: string
          description: Full CVV in the current MVP.
          example: '123'
        expiryDate:
          type: string
          description: Formatted as MM/YY.
          example: 05/29
        expiryMonth:
          type: integer
          example: 5
        expiryYear:
          type: integer
          example: 2029
        cardholderName:
          type: string
          example: Aisha Saleh
        amount:
          type: number
          format: decimal
          example: 750.0
        remainingAmount:
          type: number
          format: decimal
          example: 724.5
        useTimes:
          type: integer
          format: int32
          description: Total allowed uses. A value of `99999` represents the business default `unlimited`.
          default: 99999
          x-business-default: unlimited
          x-display-default: unlimited
          example: 99999
        usesLeft:
          type: integer
          format: int32
          description: Remaining allowed uses. For unlimited cards this starts at `99999` and is displayed as `unlimited` by clients.
          example: 99999
        unLockedAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-01T02:00:00Z'
        minutesToUnlock:
          type: integer
          format: int64
          example: 0
        hoursToUnlock:
          type: integer
          format: int64
          example: 0
        daysToUnlock:
          type: integer
          format: int64
          example: 0
        timeRemainingLeveled:
          type: string
          example: Available Now
        daysLocked:
          type: integer
          example: 0
        creatorId:
          type: integer
          example: 1
        retrieveId:
          type: integer
          description: Receiver profile id.
          example: 2
        type:
          $ref: '#/components/schemas/CardPerspectiveType'
        senderName:
          type: string
          nullable: true
          example: Mohammed Fares
    MccItem:
      type: object
      required: [code, name, group]
      properties:
        code:
          type: string
          example: '4121'
        name:
          type: string
          example: Taxis & Rideshare
        group:
          type: string
          example: Transport
...
    CardWithLogsResponse:
      type: object
      required: [card, logs]
      properties:
        card:
          $ref: '#/components/schemas/IntentWithCardResponse'
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogDto'
    PagedCardsResponse:
      type: object
      required: [total, limit, offset, items]
      properties:
        total:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/IntentWithCardResponse'
    AuditLogDto:
      type: object
      required: [id, cardId, transactionAmount, decision, createdAt]
      properties:
        id:
          type: integer
          example: 90
        cardId:
          type: integer
          nullable: true
          example: 31
        transactionAmount:
          type: number
          format: decimal
          example: 25.5
        merchantName:
          type: string
          nullable: true
          example: Riyadh Taxi
        mcc:
          type: string
          nullable: true
          example: '4121'
        decision:
          $ref: '#/components/schemas/DecisionStatus'
        reason:
          type: string
          nullable: true
          example: "Transaction declined: insufficient remaining balance."
        createdAt:
          type: string
          format: date-time
          example: '2026-05-01T02:00:00Z'
        city:
          type: string
          nullable: true
          example: Riyadh
        country:
          type: string
          nullable: true
          description: Present only if the model exposes it.
          example: SA
        externalId:
          type: string
          nullable: true
          description: Present only if the model exposes it.
          example: ext_123
        entityId:
          type: integer
          nullable: true
          description: Logical entity id, commonly the intent id.
          example: 21
        action:
          type: string
          nullable: true
          example: authorization
        occurredAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-01T02:00:00Z'
    PagedAuditLogsResponse:
      type: object
      required: [cardId, total, limit, offset, logs]
      properties:
        cardId:
          type: integer
          example: 31
        total:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogDto'
...
    CardManualFreezeStateResponse:
      type: object
      required: [cardId, isManuallyFrozen, isLockedByPendingInvoice, isSpendBlocked, previousManualFreeze]
      properties:
        cardId:
          type: integer
          example: 31
        isManuallyFrozen:
          type: boolean
          example: true
        isLockedByPendingInvoice:
          type: boolean
          example: false
        isSpendBlocked:
          type: boolean
          example: true
        previousManualFreeze:
          type: boolean
          example: false
...